requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
Werkzeug==2.0.1
python-dotenv==0.19.1
Flask-WTF==1.0.0
email-validator==1.1.3 